+++ libs/wire-api/src/Wire/API/Routes/Public/Galley.hs
@@ -318,6 +318,12 @@ data Api routes = Api
     teamFeatureStatusAppLockPut ::
       routes
         :- FeatureStatusPut 'TeamFeatureAppLock,
+    teamFeatureStatusFileSharingGet ::
+      routes
+        :- FeatureStatusGet 'TeamFeatureFileSharing,
+    teamFeatureStatusFileSharingPut ::
+      routes
+        :- FeatureStatusPut 'TeamFeatureFileSharing,
     teamFeatureStatusClassifiedDomainsGet ::
       routes
         :- FeatureStatusGet 'TeamFeatureClassifiedDomains