+++ services/galley/src/Galley/API/Teams/Features.hs
@@ -32,6 +32,8 @@ module Galley.API.Teams.Features
     getClassifiedDomainsInternal,
     getAppLockInternal,
     setAppLockInternal,
+    getFileSharingInternal,
+    setFileSharingInternal,
     DoAuth (..),
   )
 where
@@ -107,6 +109,7 @@ getAllFeatures uid tid = do
         getStatus @'Public.TeamFeatureValidateSAMLEmails getValidateSAMLEmailsInternal,
         getStatus @'Public.TeamFeatureDigitalSignatures getDigitalSignaturesInternal,
         getStatus @'Public.TeamFeatureAppLock getAppLockInternal,
+        getStatus @'Public.TeamFeatureFileSharing getFileSharingInternal,
         getStatus @'Public.TeamFeatureClassifiedDomains getClassifiedDomainsInternal
       ]
   where
@@ -212,6 +215,12 @@ setLegalholdStatusInternal tid status@(Public.tfwoStatus -> statusValue) = do
     Public.TeamFeatureEnabled -> do
       ensureNotTooLargeToActivateLegalHold tid
   TeamFeatures.setFeatureStatusNoConfig @'Public.TeamFeatureLegalHold tid status
+
+getFileSharingInternal :: TeamId -> Galley (Public.TeamFeatureStatus 'Public.TeamFeatureFileSharing)
+getFileSharingInternal = getFeatureStatusNoConfig @'Public.TeamFeatureFileSharing $ pure Public.TeamFeatureEnabled
+
+setFileSharingInternal :: TeamId -> Public.TeamFeatureStatus 'Public.TeamFeatureFileSharing -> Galley (Public.TeamFeatureStatus 'Public.TeamFeatureFileSharing)
+setFileSharingInternal = setFeatureStatusNoConfig @'Public.TeamFeatureFileSharing $ \_ _ -> pure ()
 
 getAppLockInternal :: TeamId -> Galley (Public.TeamFeatureStatus 'Public.TeamFeatureAppLock)
 getAppLockInternal tid = do